requirements.txt
openai==1.70.0
pymupdf==1.25.5
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.7.0
pyyaml==6.0.2
azure-identity
langchain==0.3.25
langchain-openai==0.3.19
langgraph==0.4.8
langgraph-checkpoint-postgres==2.0.21
neo4j-graphrag
neo4j
numpy
langchain-community
langchain-experimental
langchain-core
loguru
pyvis
pyecharts
fastapi
uvicorn
python-multipart
streamlit
requests
streamlit-echarts
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy
psycopg2-binary
pydantic[email]
minio
# Elasticsearch相关
elasticsearch[async]==8.11.0
elasticsearch-dsl==8.11.0
# Excel处理
openpyxl==3.1.2
xlrd==2.0.1
# 中文分词 - 核心依赖
jieba==0.42.1
aioboto3==15.0.0
redis
hiredis==2.2.3
langchain-ollama
werkzeug
json-repair
datasets